Renamed components, implemented some of the requested changes
