--- kura/examples/org.eclipse.kura.example.wire.bool.multiport.provider/pom.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Copyright (c) 2020 Eurotech and/or its affiliates and others
-
-     All rights reserved. This program and the accompanying materials
-     are made available under the terms of the Eclipse Public License v1.0
-     which accompanies this distribution, and is available at
-     http://www.eclipse.org/legal/epl-v10.html
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.eclipse.kura</groupId>
-        <artifactId>examples</artifactId>
-        <version>4.2.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>org.eclipse.kura.example.wire.bool.multiport.provider</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-    <packaging>eclipse-plugin</packaging>
-
-    <properties>
-       <kura.basedir>${project.basedir}/../..</kura.basedir>
-    </properties>
-
-    <build>
-      <plugins>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-eclipse-plugin</artifactId>
-          <version>2.9</version>
-          <configuration>
-            <skip>false</skip>
-          </configuration>
-        </plugin>
-          </plugins>
-    </build>
-</project>
