+++ kura/examples/org.eclipse.kura.example.wire.logic.multiport.provider/OSGI-INF/metatype/org.eclipse.kura.wire.LogicalOperators.xml
