+++ kura/examples/org.eclipse.kura.example.wire.logic.multiport.provider/build.properties
