+++ kura/examples/test/org.eclipse.kura.example.wire.logic.multiport.provider.test/build.properties
